Reworked state-viewer to show peers and use default home_dir config as near binary does
