+++ test-utils/state-viewer/Cargo.toml
@@ -12,4 +12,5 @@ near-primitives = { path = "../../core/primitives" }
 near-store = { path = "../../core/store" }
 node-runtime = { path = "../../runtime/runtime" }
 near-chain = { path = "../../chain/chain" }
+near-network = { path = "../../chain/network" }
 near = { path = "../../near" }
