+++ test-utils/state-viewer/src/main.rs
@@ -1,11 +1,11 @@
 use std::convert::TryFrom;
-use std::path::PathBuf;
+use std::path::Path;
 
-use clap::{App, Arg};
+use clap::{App, Arg, SubCommand};
 
-use near::config::GenesisConfig;
-use near::{get_store_path, NightshadeRuntime};
+use near::{get_default_home, get_store_path, load_config, NightshadeRuntime};
 use near_chain::{ChainStore, ChainStoreAccess};
+use near_network::peer_store::PeerStore;
 use near_primitives::account::{AccessKey, Account};
 use near_primitives::crypto::signature::PublicKey;
 use near_primitives::hash::hash;
@@ -15,8 +15,6 @@ use near_primitives::transaction::Callback;
 use near_primitives::utils::col;
 use near_store::{create_store, DBValue, TrieIterator};
 use node_runtime::ext::ACCOUNT_DATA_SEPARATOR;
-
-const DEFAULT_BASE_PATH: &str = "";
 
 fn to_printable(blob: &[u8]) -> String {
     if blob.len() > 60 {
@@ -82,51 +80,47 @@ fn print_state_entry(key: Vec<u8>, value: DBValue) {
 
 fn main() {
     init_integration_logger();
+
+    let default_home = get_default_home();
     let matches = App::new("state-viewer")
-        .args(&[
-            Arg::with_name("base_path")
-                .short("d")
-                .long("base-path")
-                .value_name("PATH")
-                .help("Specify a base path for persisted files.")
-                .default_value(DEFAULT_BASE_PATH)
+        .arg(
+            Arg::with_name("home")
+                .long("home")
+                .default_value(&default_home)
+                .help("Directory for config and data (default \"~/.near\")")
                 .takes_value(true),
-            Arg::with_name("genesis")
-                .short("c")
-                .long("chain-spec-file")
-                .value_name("CHAIN_SPEC")
-                .help("Specify a file location to read a custom chain spec.")
-                .takes_value(true),
-            Arg::with_name("devnet")
-                .long("devnet")
-                .help("Run with DevNet validator configuration (single alice.near validator)")
-                .takes_value(false),
-        ])
+        )
+        .subcommand(SubCommand::with_name("peers"))
+        .subcommand(SubCommand::with_name("state"))
         .get_matches();
-    let base_path = matches.value_of("base_path").map(PathBuf::from).unwrap();
-    let genesis_config = if matches.is_present("devnet") {
-        GenesisConfig::legacy_test(vec!["alice.near", "bob.near", "carol.near"], 1)
-    } else {
-        let chain_spec_path = matches.value_of("chain_spec_file").map(PathBuf::from);
-        if let Some(path) = chain_spec_path {
-            GenesisConfig::from_file(&path)
-        } else {
-            GenesisConfig::testing_spec(3, 2)
+
+    let home_dir = matches.value_of("home").map(|dir| Path::new(dir)).unwrap();
+    let near_config = load_config(home_dir);
+
+    let store = create_store(&get_store_path(&home_dir));
+
+    match matches.subcommand() {
+        ("peers", Some(_args)) => {
+            let peer_store = PeerStore::new(store.clone(), &vec![]).unwrap();
+            for (peer_id, peer_info) in peer_store.iter() {
+                println!("{} {:?}", peer_id, peer_info);
+            }
         }
-    };
+        ("state", Some(_args)) => {
+            let mut chain_store = ChainStore::new(store.clone());
 
-    let store = create_store(&get_store_path(&base_path));
-    let mut chain_store = ChainStore::new(store.clone());
+            let runtime = NightshadeRuntime::new(&home_dir, store, near_config.genesis_config);
+            let head = chain_store.head().unwrap();
+            let last_header = chain_store.get_block_header(&head.last_block_hash).unwrap().clone();
+            let state_root = chain_store.get_post_state_root(&head.last_block_hash).unwrap();
+            let trie = TrieIterator::new(&runtime.trie, state_root).unwrap();
 
-    let runtime = NightshadeRuntime::new(&base_path, store, genesis_config);
-    let head = chain_store.head().unwrap();
-    let last_header = chain_store.get_block_header(&head.last_block_hash).unwrap().clone();
-    let state_root = chain_store.get_post_state_root(&head.last_block_hash).unwrap();
-    let trie = TrieIterator::new(&runtime.trie, state_root).unwrap();
-
-    println!("Storage root is {}, block height is {}", state_root, last_header.height);
-    for item in trie {
-        let (key, value) = item.unwrap();
-        print_state_entry(key, value);
+            println!("Storage root is {}, block height is {}", state_root, last_header.height);
+            for item in trie {
+                let (key, value) = item.unwrap();
+                print_state_entry(key, value);
+            }
+        }
+        (_, _) => unreachable!(),
     }
 }
